projects/dsb/lightstream/pyproject.toml
[project]
name = "lightstream"
version = "0.1.0"
description = "PoW DDos protection"
requires-python = "==3.13.2"
dependencies = [
  "argon2-cffi==23.1.0",
  "cryptography==44.0.2",
  "fastapi==0.115.12",
  "pyjwt==2.10.1",
]
